Curl fallback mechanism
It's now possible to use curl when querying the api. You can do so by setting the third initialisation parameter to true (default false).

Curl has to be available on the system, and added to PATH. 

The CliWrap nuget is used to execute curl, either add a reference in your project or make sure the required dll's are added.